% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{download_bahm}
\alias{download_bahm}
\title{Download an instance of bahm to your machine}
\usage{
download_bahm(path = getwd())
}
\arguments{
\item{path}{The directory to download bahm to}
}
\description{
Download an instance of bahm to your machine
}
\examples{
\dontrun{
    download_bahm()
}
}
