% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run_bahm}
\alias{run_bahm}
\title{Run an instance of bahm on your machine}
\usage{
run_bahm(path = getwd())
}
\arguments{
\item{path}{The directory to run bahm from}
}
\description{
Run an instance of bahm on your machine
}
\examples{
\dontrun{
    run_bahm()
}
}
